all done

+++ data/migrations/01-make_cars_table.js
@@ -5,7 +5,7 @@ exports.up = function(knex) {
         table.text("vin",128).unique().notNullable()
         table.text("make", 128).notNullable()
         table.text("model", 128).notNullable()
-        table.decimal("mileage").notNullable()
+        table.integer("mileage").notNullable()
         table.text("title", 128)
         table.text("transmission")
     })
